move reader into module
